Extract DurableTask.Core shims to common Worker package
