+++ src/Worker/Grpc/GrpcExtensions.cs
@@ -1,16 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using System.Threading;
 using Grpc.Core;
-using Microsoft.DurableTask.Converters;
-using Microsoft.DurableTask.Grpc;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
 namespace Microsoft.DurableTask;
